Fixed issue with crashing CI tests
- Added --headless=true to HAGrid to disable the frontend
- Refactored integration tests into module groupings
- Shutdown frontend after frontend tests to save on CI RAM
- Fixed numpy warning with type np.float
- Removed gzipping from the docker image save / load to speed it up
