+++ packages/grid/docker-compose.override.yml
@@ -129,6 +129,8 @@ services:
     network_mode: service:proxy
 
   frontend:
+    profiles:
+      - frontend
     build:
       context: ./frontend
       dockerfile: frontend.dockerfile
