+++ tests/integration/conftest.py
@@ -6,6 +6,7 @@ from typing import Callable
 from typing import List
 
 # third party
+import _pytest
 import pytest
 
 # syft absolute
@@ -36,3 +37,13 @@ def get_clients() -> Callable[[int], List[Any]]:
         return clients[:nr_clients]
 
     return _helper_get_clients
+
+
+def pytest_configure(config: _pytest.config.Config) -> None:
+    config.addinivalue_line("markers", "general: general integration tests")
+    config.addinivalue_line("markers", "frontend: frontend integration tests")
+    config.addinivalue_line("markers", "smpc: smpc integration tests")
+    config.addinivalue_line("markers", "network: network integration tests")
+    config.addinivalue_line("markers", "k8s: kubernetes integration tests")
+    config.addinivalue_line("markers", "e2e: end-to-end integration tests")
+    config.addinivalue_line("markers", "security: security integration tests")
