+++ tests/integration/e2e/trade_demo_smpc_adp_test.py
@@ -17,7 +17,7 @@ from syft.core.adp.entity import Entity
 sy.logger.remove()
 
 
-ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../"))
+ROOT_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../../"))
 print("ROOT_DIR", ROOT_DIR)
 
 
@@ -34,40 +34,30 @@ def get_user_details(unique_email: str) -> Dict[str, Any]:
     }
 
 
-@pytest.mark.integration
+@pytest.mark.e2e
 def test_end_to_end_smpc_adp_trade_demo() -> None:
-    # stdlib
-    from inspect import currentframe
-    from inspect import getframeinfo
-
-    print("running test_end_to_end_smpc_adp_trade_demo")
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
     # make a unique email so we can run the test isolated
     unique_email = f"{uuid.uuid4()}@caltech.edu"
 
     # Canada
     ca_root = sy.login(email="info@openmined.org", password="changethis", port=9082)
     ca_data = load_data(csv_file="ca - feb 2021.csv")
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     # NOTE: casting this tensor as np.int32 is REALLY IMPORTANT
     canada_trade = (
         (np.array(list(ca_data["Trade Value (US$)"])) / 100000)[0:10]
     ).astype(np.int32)
     trade_partners = ((list(ca_data["Partner"])))[0:10]
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     entities = list()
     for i in range(len(trade_partners)):
         entities.append(Entity(name=trade_partners[i]))
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     sampled_canada_dataset = sy.Tensor(canada_trade)
     sampled_canada_dataset.public_shape = sampled_canada_dataset.shape
     sampled_canada_dataset = sampled_canada_dataset.private(
         0, 3, entities=entities[0]
     ).tag("trade_flow")
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     # load dataset
     ca_root.load_dataset(
@@ -79,13 +69,11 @@ def test_end_to_end_smpc_adp_trade_demo() -> None:
         ),
         skip_checks=True,
     )
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     assert len(ca_root.datasets) > 0
 
     # create user
     ca_root.users.create(**get_user_details(unique_email=unique_email))
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     # Italy
     it_root = sy.login(email="info@openmined.org", password="changethis", port=9083)
@@ -95,12 +83,10 @@ def test_end_to_end_smpc_adp_trade_demo() -> None:
         np.int32
     )
     trade_partners = ((list(it_data["Partner"])))[0:10]
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     entities = list()
     for i in range(len(trade_partners)):
         entities.append(Entity(name="Other Asia, nes"))
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     # Upload a private dataset to the Domain object, as the root owner
     sampled_italy_dataset = sy.Tensor(data_batch)
@@ -108,7 +94,6 @@ def test_end_to_end_smpc_adp_trade_demo() -> None:
     sampled_italy_dataset = sampled_italy_dataset.private(
         0, 3, entities=entities[0]
     ).tag("trade_flow")
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     it_root.load_dataset(
         assets={"Italy Trade": sampled_italy_dataset},
@@ -119,51 +104,41 @@ def test_end_to_end_smpc_adp_trade_demo() -> None:
         ),
         skip_checks=True,
     )
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     assert len(it_root.datasets) > 0
 
     # create user
     it_root.users.create(**get_user_details(unique_email=unique_email))
 
-    print("what user are we requesting and not approving", unique_email)
-
     # Data Scientist
     ca = sy.login(email=unique_email, password="bazinga", port=9082)
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     ca.request_budget(eps=200, reason="increase budget!")
     it = sy.login(email=unique_email, password="bazinga", port=9083)
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     it.request_budget(eps=200, reason="increase budget!")
 
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
     time.sleep(10)
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     # until we fix the code this just accepts all requests in case it gets the
     # wrong one
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     for req in ca_root.requests:
         req.accept()
 
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
     for req in it_root.requests:
         req.accept()
 
     # ca_root.requests[0].accept()
     # it_root.requests[0].accept()
 
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
     time.sleep(10)
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     assert round(ca.privacy_budget) == 210
     assert round(it.privacy_budget) == 210
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     ca_data = ca.datasets[-1]["Canada Trade"]
     it_data = it.datasets[-1]["Italy Trade"]
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     """
     Cutter: Every magic trick consists of three parts, or acts.
@@ -176,11 +151,10 @@ def test_end_to_end_smpc_adp_trade_demo() -> None:
     """
     # the pledge 🦜
     print("running the pledge 🦜")
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     result = ca_data + it_data
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     result.block_with_timeout(40)
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     """
     Cutter: The second act is called "The Turn". The mathemagician takes the ordinary
@@ -189,9 +163,8 @@ def test_end_to_end_smpc_adp_trade_demo() -> None:
     """
     # the turn 🕳
     print("running the turn 🕳")
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     public_result = result.publish(sigma=2)
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     """
     Cutter: But you wouldn't clap yet. Because making information disappear isn't enough;
@@ -201,25 +174,21 @@ def test_end_to_end_smpc_adp_trade_demo() -> None:
     # the prestige 🎩
     print("running the prestige 🎩")
     # time.sleep(40)  # TODO: should modify after implementing polling .get()
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     public_result.block_with_timeout(40)
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     sycure_result = public_result.get()
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
+
     print("sycure_result", sycure_result)
     print("after ca", ca.privacy_budget)
     print("after it", it.privacy_budget)
 
     assert len(sycure_result) == 10
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
     assert sum(sycure_result) > -100
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
     assert sum(sycure_result) < 100
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
 
     assert ca.privacy_budget < 210
     assert ca.privacy_budget > 10
     assert it.privacy_budget < 210
     assert it.privacy_budget > 10
     assert ca.privacy_budget == it.privacy_budget
-    print("test_end_to_end_smpc_adp_trade_demo", getframeinfo(currentframe()).lineno)
