+++ tests/integration/security/check_vpn_firewall_test.py
@@ -2,6 +2,9 @@
 import re
 import subprocess
 
+# third party
+import pytest
+
 
 def docker_network_connect(direction: str = "connect") -> None:
     # this connects all the tailscale containers to the other docker compose project
@@ -25,10 +28,12 @@ def docker_network_connect(direction: str = "connect") -> None:
                 print(f"Exception running: {cmd}. {e}")
 
 
+@pytest.mark.security
 def test_create_overlay_networks_docker() -> None:
     docker_network_connect(direction="connect")
 
 
+@pytest.mark.security
 def test_vpn_scan() -> None:
     # the tailscale container is currently the same so we can get away with a
     # single external scan
@@ -61,5 +66,6 @@ def test_vpn_scan() -> None:
             assert len(lines) == 0
 
 
+@pytest.mark.security
 def test_remove_overlay_networks_docker() -> None:
     docker_network_connect(direction="disconnect")
