+++ tox.ini
@@ -109,7 +109,7 @@ commands =
     bash -c "docker volume rm test_network_1_headscale-data --force || true"
     bash -c 'HAGRID_ART=false hagrid launch test_network_1 network to docker:9081 --tail=false'
     bash -c 'HAGRID_ART=false hagrid launch test_domain_1 domain to docker:9082 --tail=false'
-    bash -c 'HAGRID_ART=false hagrid launch test_domain_2 domain to docker:9083 --tail=false'
+    bash -c 'HAGRID_ART=false hagrid launch test_domain_2 domain to docker:9083 --tail=false --headless=true'
     docker ps
     bash -c '(docker logs test_domain_1-frontend-1 -f &) | grep -q "event - compiled successfully" || true'
     bash -c '(docker logs test_network_1-frontend-1 -f &) | grep -q "event - compiled successfully" || true'
@@ -118,13 +118,25 @@ commands =
     bash -c '(docker logs test_network_1-backend_stream-1 -f &) | grep -q "Application startup complete" || true'
 
     sleep 5
-    ; connect each stack to the VPN
-    pytest tests/vpn_connect -p no:randomly
-    pytest tests/association_requests -p no:randomly
-    pytest tests/integration -vvvv -k "not k8s and not test_end_to_end_smpc_adp_trade_demo" --ignore=tests/integration/smpc -p no:randomly -p no:benchmark --co
-    pytest tests/integration -vvvv -k "not k8s and not test_end_to_end_smpc_adp_trade_demo" --ignore=tests/integration/smpc -p no:randomly -p no:benchmark -o log_cli=True --capture=no
-    ; pytest tests/integration -vvvv -k "test_end_to_end_smpc_adp_trade_demo" --ignore=tests/integration/smpc -p no:randomly -p no:benchmark -o log_cli=True --capture=no
-    pytest tests/scan_network -p no:randomly
+
+    pytest tests/integration -m frontend -p no:randomly --co
+    pytest tests/integration -m frontend -vvvv -p no:randomly -p no:benchmark -o log_cli=True --capture=no
+
+    bash -c 'docker stop test_network_1-frontend-1 || true'
+    bash -c 'docker stop test_domain_1-frontend-1 || true'
+
+    pytest tests/integration -m network -p no:randomly --co
+    pytest tests/integration -m network -vvvv -p no:randomly -p no:benchmark -o log_cli=True --capture=no
+
+    pytest tests/integration -m general -p no:randomly --co
+    pytest tests/integration -m general -vvvv -p no:randomly -p no:benchmark -o log_cli=True --capture=no
+
+    pytest tests/integration -m e2e -p no:randomly --co
+    pytest tests/integration -m e2e -vvvv -p no:randomly -p no:benchmark -o log_cli=True --capture=no
+
+    pytest tests/integration -m security -p no:randomly --co
+    pytest tests/integration -m security -vvvv -p no:randomly -p no:benchmark -o log_cli=True --capture=no
+
     bash -c 'HAGRID_ART=false hagrid land test_network_1'
     bash -c 'HAGRID_ART=false hagrid land test_domain_1'
     bash -c 'HAGRID_ART=false hagrid land test_domain_2'
@@ -149,20 +161,21 @@ commands =
     bash -c "docker volume rm test_domain_1_tailscale-data --force || true"
     bash -c "docker volume rm test_domain_2_tailscale-data --force || true"
     bash -c "docker volume rm test_domain_3_tailscale-data --force || true"
-    bash -c 'HAGRID_ART=false hagrid launch test_domain_1 domain to docker:9082 --tail=false'
-    bash -c 'HAGRID_ART=false hagrid launch test_domain_2 domain to docker:9083 --tail=false'
-    bash -c 'HAGRID_ART=false hagrid launch test_domain_3 domain to docker:9084 --tail=false'
+    bash -c 'HAGRID_ART=false hagrid launch test_domain_1 domain to docker:9082 --tail=false --headless=true'
+    bash -c 'HAGRID_ART=false hagrid launch test_domain_2 domain to docker:9083 --tail=false --headless=true'
+    bash -c 'HAGRID_ART=false hagrid launch test_domain_3 domain to docker:9084 --tail=false --headless=true'
 
     docker ps
     bash -c '(docker logs test_domain_1-backend_stream-1 -f &) | grep -q "Application startup complete" || true'
     bash -c '(docker logs test_domain_2-backend_stream-1 -f &) | grep -q "Application startup complete" || true'
     bash -c '(docker logs test_domain_3-backend_stream-1 -f &) | grep -q "Application startup complete" || true'
 
-
     sleep 5
     ; perform spmc test
 
-    pytest tests/integration/smpc -vvvv -p no:randomly
+    pytest tests/integration -m smpc -p no:randomly --co
+    pytest tests/integration -m smpc -vvvv -p no:randomly -p no:benchmark -o log_cli=True --capture=no
+
     bash -c 'HAGRID_ART=false hagrid land test_domain_1'
     bash -c 'HAGRID_ART=false hagrid land test_domain_2'
     bash -c 'HAGRID_ART=false hagrid land test_domain_3'
@@ -190,7 +203,9 @@ commands =
     devspace use namespace openmined
     bash -c 'cd packages/grid && devspace deploy -b -p domain'
     bash -c 'while ! (kubectl get ingress &) | grep -Eq "stack-ingress.*[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}"; do echo "waiting for ingress" && sleep 1; done'
-    bash -c 'TEST_DOMAIN_IP=`minikube ip` TEST_DOMAIN_PORT=80 pytest tests/integration/k8s_test.py -vvvv -p no:randomly'
+
+    pytest -m integration.k8s -p no:randomly --co
+    bash -c 'TEST_DOMAIN_IP=`minikube ip` TEST_DOMAIN_PORT=80 pytest -m integration.k8s -vvvv -p no:randomly -p no:benchmark -o log_cli=True --capture=no'
 
 [testenv:grid.test.backend]
 description = Tests for Grid Backend
